convert editprofile to cbv and move ProfileListView
