+++ userena/views/__init__.py
@@ -1,6 +1,5 @@
 from .signup import SignupView
 from .signin import SigninView
-from .profile import ProfileDetailView
+from .profile import ProfileDetailView, ProfileListView, ProfileEditView
 from .views import signout, direct_to_user_template, activate, activate_retry, email_change, direct_to_user_template
 from .views import direct_to_user_template, email_confirm, disabled_account, password_change, direct_to_user_template
-from .views import profile_edit, ProfileListView
