+++ userena/views/views.py
@@ -6,10 +6,9 @@ from django.core.urlresolvers import reverse
 from django.http import Http404
 from django.shortcuts import redirect, get_object_or_404
 from django.views.generic import TemplateView
-from django.views.generic.list import ListView
 from django.utils.translation import ugettext as _
 
-from userena.forms import (ChangeEmailForm, EditProfileForm)
+from userena.forms import (ChangeEmailForm)
 from userena.models import UserenaSignup
 from userena.decorators import secure_required
 from userena.utils import (get_profile_model, get_user_model,
@@ -33,39 +32,6 @@ class ExtraContextTemplateView(TemplateView):
 
     # this view is used in POST requests, e.g. signup when the form is not valid
     post = TemplateView.get
-
-class ProfileListView(ListView):
-    """ Lists all profiles """
-    context_object_name='profile_list'
-    page=1
-    paginate_by=50
-    template_name=userena_settings.USERENA_PROFILE_LIST_TEMPLATE
-    extra_context=None
-
-    def get_context_data(self, **kwargs):
-        # Call the base implementation first to get a context
-        context = super(ProfileListView, self).get_context_data(**kwargs)
-        try:
-            page = int(self.request.GET.get('page', None))
-        except (TypeError, ValueError):
-            page = self.page
-
-        if userena_settings.USERENA_DISABLE_PROFILE_LIST \
-           and not self.request.user.is_staff:
-            raise Http404
-
-        if not self.extra_context: self.extra_context = dict()
-
-        context['page'] = page
-        context['paginate_by'] = self.paginate_by
-        context['extra_context'] = self.extra_context
-
-        return context
-
-    def get_queryset(self):
-        profile_model = get_profile_model()
-        queryset = profile_model.objects.get_visible_profiles(self.request.user).select_related()
-        return queryset
 
 
 @secure_required
@@ -465,85 +431,6 @@ def password_change(request, username, template_name='userena/password_form.html
     extra_context['profile'] = get_user_profile(user=user)
     return ExtraContextTemplateView.as_view(template_name=template_name,
                                             extra_context=extra_context)(request)
-@secure_required
-@permission_required_or_403('change_profile', (get_profile_model(), 'user__username', 'username'))
-def profile_edit(request, username, edit_profile_form=EditProfileForm,
-                 template_name='userena/profile_form.html', success_url=None,
-                 extra_context=None, **kwargs):
-    """
-    Edit profile.
-
-    Edits a profile selected by the supplied username. First checks
-    permissions if the user is allowed to edit this profile, if denied will
-    show a 404. When the profile is successfully edited will redirect to
-    ``success_url``.
-
-    :param username:
-        Username of the user which profile should be edited.
-
-    :param edit_profile_form:
-
-        Form that is used to edit the profile. The :func:`EditProfileForm.save`
-        method of this form will be called when the form
-        :func:`EditProfileForm.is_valid`.  Defaults to :class:`EditProfileForm`
-        from userena.
-
-    :param template_name:
-        String of the template that is used to render this view. Defaults to
-        ``userena/edit_profile_form.html``.
-
-    :param success_url:
-        Named URL which will be passed on to a django ``reverse`` function after
-        the form is successfully saved. Defaults to the ``userena_detail`` url.
-
-    :param extra_context:
-        Dictionary containing variables that are passed on to the
-        ``template_name`` template.  ``form`` key will always be the form used
-        to edit the profile, and the ``profile`` key is always the edited
-        profile.
-
-    **Context**
-
-    ``form``
-        Form that is used to alter the profile.
-
-    ``profile``
-        Instance of the ``Profile`` that is edited.
-
-    """
-    user = get_object_or_404(get_user_model(), username__iexact=username)
-
-    profile = get_user_profile(user=user)
-
-    user_initial = {'first_name': user.first_name,
-                    'last_name': user.last_name}
-
-    form = edit_profile_form(instance=profile, initial=user_initial)
-
-    if request.method == 'POST':
-        form = edit_profile_form(request.POST, request.FILES, instance=profile,
-                                 initial=user_initial)
-
-        if form.is_valid():
-            profile = form.save()
-
-            if userena_settings.USERENA_USE_MESSAGES:
-                messages.success(request, _('Your profile has been updated.'),
-                                 fail_silently=True)
-
-            if success_url:
-                # Send a signal that the profile has changed
-                userena_signals.profile_change.send(sender=None,
-                                                    user=user)
-                redirect_to = success_url
-            else: redirect_to = reverse('userena_profile_detail', kwargs={'username': username})
-            return redirect(redirect_to)
-
-    if not extra_context: extra_context = dict()
-    extra_context['form'] = form
-    extra_context['profile'] = profile
-    return ExtraContextTemplateView.as_view(template_name=template_name,
-                                            extra_context=extra_context)(request)
 
 
 def profile_list(request, page=1, template_name='userena/profile_list.html',
@@ -602,6 +489,7 @@ def profile_list(request, page=1, template_name='userena/profile_list.html',
     queryset = profile_model.objects.get_visible_profiles(request.user)
 
     if not extra_context: extra_context = dict()
+    from .profile import ProfileListView
     return ProfileListView.as_view(queryset=queryset,
                                    paginate_by=paginate_by,
                                    page=page,
